apps/workers/src/processors/image_processor.py
"""Image processing operations for media preprocessing."""

import logging
from typing import Dict, Any, Optional
from PIL import Image, ImageEnhance, ImageFilter
import numpy as np

logger = logging.getLogger(__name__)

class ImageProcessor:
    """Handles basic image processing operations."""
    
    def __init__(self):
        self.supported_formats = ['jpeg', 'jpg', 'png', 'webp']
    
    async def resize(self, config: Dict[str, Any]) -> Dict[str, Any]:
        """Resize image according to configuration."""
        try:
            # Extract configuration
            width = config.get('width')
            height = config.get('height')
            max_width = config.get('max_width')
            max_height = config.get('max_height')
            maintain_aspect_ratio = config.get('maintain_aspect_ratio', True)
            quality = config.get('quality', 85)
            
            # Simulate image processing
            logger.info(f"Resizing image: width={width}, height={height}, max_width={max_width}, max_height={max_height}")
            
            # In real implementation, would load actual image and resize
            # For now, return mock result
            return {
                "status": "success",
                "original_size": {"width": 1920, "height": 1080},
                "new_size": {"width": width or max_width or 800, "height": height or max_height or 600},
                "maintained_aspect_ratio": maintain_aspect_ratio,
                "quality": quality
            }
            
        except Exception as e:
            logger.error(f"Resize operation failed: {str(e)}")
            return {"status": "error", "message": str(e)}
    
    async def crop(self, config: Dict[str, Any]) -> Dict[str, Any]:
        """Crop image according to configuration."""
        try:
            # Extract configuration
            x = config.get('x')
            y = config.get('y')
            width = config.get('width')
            height = config.get('height')
            center_crop = config.get('center_crop', False)
            aspect_ratio = config.get('aspect_ratio')
            
            logger.info(f"Cropping image: x={x}, y={y}, width={width}, height={height}, center_crop={center_crop}")
            
            # In real implementation, would load actual image and crop
            # For now, return mock result
            return {
                "status": "success",
                "crop_box": {"x": x or 0, "y": y or 0, "width": width or 800, "height": height or 600},
                "center_crop": center_crop,
                "aspect_ratio": aspect_ratio
            }
            
        except Exception as e:
            logger.error(f"Crop operation failed: {str(e)}")
            return {"status": "error", "message": str(e)}
    
    async def normalize(self, config: Dict[str, Any]) -> Dict[str, Any]:
        """Normalize image pixel values."""
        try:
            # Extract configuration
            mean = config.get('mean', [0.485, 0.456, 0.406])
            std = config.get('std', [0.229, 0.224, 0.225])
            scale = config.get('scale', 1.0 / 255.0)
            
            logger.info(f"Normalizing image: mean={mean}, std={std}, scale={scale}")
            
            # In real implementation, would load actual image and normalize
            # For now, return mock result
            return {
                "status": "success",
                "mean": mean,
                "std": std,
                "scale": scale,
                "normalized": True
            }
            
        except Exception as e:
            logger.error(f"Normalize operation failed: {str(e)}")
            return {"status": "error", "message": str(e)}
    
    async def enhance(self, config: Dict[str, Any]) -> Dict[str, Any]:
        """Enhance image with various adjustments."""
        try:
            # Extract configuration
            brightness = config.get('brightness', 1.0)
            contrast = config.get('contrast', 1.0)
            saturation = config.get('saturation', 1.0)
            sharpness = config.get('sharpness', 1.0)
            
            logger.info(f"Enhancing image: brightness={brightness}, contrast={contrast}, saturation={saturation}, sharpness={sharpness}")
            
            # In real implementation, would load actual image and apply enhancements
            # For now, return mock result
            return {
                "status": "success",
                "enhancements": {
                    "brightness": brightness,
                    "contrast": contrast,
                    "saturation": saturation,
                    "sharpness": sharpness
                }
            }
            
        except Exception as e:
            logger.error(f"Enhance operation failed: {str(e)}")
            return {"status": "error", "message": str(e)}
    
    async def create_thumbnail(self, config: Dict[str, Any]) -> Dict[str, Any]:
        """Create thumbnail of image."""
        try:
            # Extract configuration
            width = config.get('width', 150)
            height = config.get('height', 150)
            quality = config.get('quality', 85)
            
            logger.info(f"Creating thumbnail: width={width}, height={height}, quality={quality}")
            
            # In real implementation, would load actual image and create thumbnail
            # For now, return mock result
            return {
                "status": "success",
                "thumbnail_size": {"width": width, "height": height},
                "quality": quality,
                "thumbnail_url": f"/thumbnails/{width}x{height}/image.jpg"
            }
            
        except Exception as e:
            logger.error(f"Thumbnail creation failed: {str(e)}")
            return {"status": "error", "message": str(e)}
    
    async def get_image_info(self, image_path: str) -> Dict[str, Any]:
        """Get basic information about an image."""
        try:
            # In real implementation, would load actual image and get info
            # For now, return mock result
            return {
                "format": "JPEG",
                "mode": "RGB",
                "size": {"width": 1920, "height": 1080},
                "file_size": 1024000,
                "dpi": (72, 72),
                "has_transparency": False
            }
            
        except Exception as e:
            logger.error(f"Get image info failed: {str(e)}")
            return {"status": "error", "message": str(e)}
